Move common app parts to shared folder - part I
