+++ frontend/src/app/modules/nav/nav.component.ts
@@ -2,7 +2,7 @@ import {Component, EventEmitter, OnInit, Output} from '@angular/core';
 import {Observable} from 'rxjs';
 
 import {AuthService} from '../auth/auth.service';
-import {BreakpointService} from '../../services/breakpoint.service';
+import {BreakpointService} from '../../shared/services/breakpoint.service';
 
 @Component({
   selector: 'app-nav',
